pyproject.toml
[tool.poetry]
name = "project1"
version = "0.1.0"
description = ""
authors = ["Valery Shcherbakov <val032606@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.12"
selenium = "^4.31.0"
pyperclip = "^1.9.0"
allure-pytest = "^2.14.0"
python-dotenv = "^1.1.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
